Add files via upload
webbrowser :Comes with Python and opens a browser to a specific page.

Requests: Downloads files and web pages from the Internet.

Beautiful Soup: Parses HTML, the format that web pages are written in.

Selenium: Launches and controls a web browser. Selenium is able to
fill in forms and simulate mouse clicks in this browser.